Fix stale test class namespaces left behind by Neo4j plugin move
The PR that moved Neo4j-bound code into src/GraphDatabasePlugins/Neo4j/
(and the corresponding tests under tests/phpunit/GraphDatabasePlugins/Neo4j/)
updated the source-side namespaces but left six test files with namespace
declarations that no longer match their PSR-4 path. Composer can therefore
not autoload these test classes by FQN.

- CompositeCypherQueryValidator{,Integration}Test: kept the pre-move
  ...Tests\Application prefix instead of picking up the new
  ...Tests\GraphDatabasePlugins\Neo4j\Application prefix.
- Explain/KeywordCypherQueryValidatorTest: live under Application/ but
  declared ...\Persistence.
- Neo4jResultNormalizerTest: lives under Persistence/ but declared
  ...\Application (mirroring the source move from Application/ to
  Persistence/).
- QueryExceptionTest: lives under Application/Exception/ but declared
  the parent ...\Application namespace, missing the Exception segment.

Source-side namespaces are already correct; only the test side is touched
here. No production behavior changes.
